api/requirements-test.txt
pytest==7.2.2
attrs==23.2.0
tavern==2.10.0
pytest-asyncio==0.21.1
httpx==0.24.1
Pillow==11.1.0